third_party/blink/web_tests/external/wpt/mediacapture-streams/MediaDevices-enumerateDevices-returned-objects.https-expected.txt

This is a testharness.js-based test.
FAIL enumerateDevices returns expected mostly empty objects in case device-info permission is not granted assert_equals: deviceId is empty before capture expected "" but got "default"
PASS enumerateDevices returns expected objects in case device-info permission is granted
Harness: the test ran to completion.
